+++ pydantic/env_settings.py
@@ -2,7 +2,7 @@ import json
 import os
 from typing import Dict, Optional, cast
 
-from .main import BaseModel
+from .main import BaseModel, Extra
 
 
 class SettingsError(ValueError):
@@ -58,6 +58,6 @@ class BaseSettings(BaseModel):
     class Config:
         env_prefix = 'APP_'
         validate_all = True
-        ignore_extra = False
+        extra = Extra.forbid
         arbitrary_types_allowed = True
         case_insensitive = False