+++ pydantic/fields.py
@@ -1,13 +1,13 @@
 import warnings
 from enum import IntEnum
-from typing import TYPE_CHECKING, Any, Callable, Dict, List, Mapping, Optional, Pattern, Set, Tuple, Type, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Mapping, Optional, Pattern, Set, Tuple, Type, Union
 
 from . import errors as errors_
 from .class_validators import Validator, ValidatorSignature, get_validator_signature
 from .error_wrappers import ErrorWrapper
 from .types import Json, JsonWrapper
-from .utils import ForwardRef, display_as_type, lenient_issubclass, list_like
-from .validators import NoneType, dict_validator, find_validators, is_none_validator
+from .utils import Callable, ForwardRef, display_as_type, lenient_issubclass, list_like
+from .validators import NoneType, dict_validator, find_validators
 
 Required: Any = Ellipsis
 
@@ -149,6 +149,8 @@ class Field:
         origin = getattr(self.type_, '__origin__', None)
         if origin is None:
             # field is not "typing" object eg. Union, Dict, List etc.
+            return
+        if origin is Callable:
             return
         if origin is Union:
             types_ = []
@@ -360,7 +362,7 @@ class Field:
         """
         False if this is a simple field just allowing None as used in Unions/Optional.
         """
-        return len(self.validators) != 1 or self.validators[0][1] != is_none_validator
+        return self.type_ != NoneType  # type: ignore
 
     def is_complex(self):
         """