+++ pydantic/schema.py
@@ -1,3 +1,4 @@
+import warnings
 from datetime import date, datetime, time, timedelta
 from decimal import Decimal
 from enum import Enum
@@ -29,7 +30,7 @@ from .types import (
     conint,
     constr,
 )
-from .utils import clean_docstring, lenient_issubclass
+from .utils import clean_docstring, is_callable_type, lenient_issubclass
 
 __all__ = [
     'Schema',
@@ -72,6 +73,8 @@ class Schema:
       will have an ``exclusiveMaximum`` validation keyword
     :param le: only applies to numbers, requires the field to be "less than or equal to". The
       schema will have a ``maximum`` validation keyword
+    :param multiple_of: only applies to numbers, requires the field to be "a multiple of". The
+      schema will have a ``multipleOf`` validation keyword
     :param min_length: only applies to strings, requires the field to have a minimum length. The
       schema will have a ``maximum`` validation keyword
     :param max_length: only applies to strings, requires the field to have a maximum length. The
@@ -90,6 +93,7 @@ class Schema:
         'ge',
         'lt',
         'le',
+        'multiple_of',
         'min_length',
         'max_length',
         'regex',
@@ -107,6 +111,7 @@ class Schema:
         ge: float = None,
         lt: float = None,
         le: float = None,
+        multiple_of: float = None,
         min_length: int = None,
         max_length: int = None,
         regex: str = None,
@@ -121,6 +126,7 @@ class Schema:
         self.ge = ge
         self.lt = lt
         self.le = le
+        self.multiple_of = multiple_of
         self.min_length = min_length
         self.max_length = max_length
         self.regex = regex
@@ -268,6 +274,7 @@ _numeric_types_attrs: Tuple[Tuple[str, Union[type, Tuple[type, ...]], str], ...]
     ('lt', numeric_types, 'exclusiveMaximum'),
     ('ge', numeric_types, 'minimum'),
     ('le', numeric_types, 'maximum'),
+    ('multiple_of', numeric_types, 'multipleOf'),
 )
 
 
@@ -500,9 +507,13 @@ def model_type_schema(
     required = []
     definitions: Dict[str, Any] = {}
     for k, f in model.__fields__.items():
-        f_schema, f_definitions = field_schema(
-            f, by_alias=by_alias, model_name_map=model_name_map, ref_prefix=ref_prefix
-        )
+        try:
+            f_schema, f_definitions = field_schema(
+                f, by_alias=by_alias, model_name_map=model_name_map, ref_prefix=ref_prefix
+            )
+        except SkipField as skip:
+            warnings.warn(skip.message, UserWarning)
+            continue
         definitions.update(f_definitions)
         if by_alias:
             properties[f.alias] = f_schema
@@ -566,6 +577,7 @@ validation_attribute_to_schema_keyword = {
     'lt': 'exclusiveMaximum',
     'ge': 'minimum',
     'le': 'maximum',
+    'multiple_of': 'multipleOf',
 }
 
 # Order is important, subclasses of str must go before str, etc
@@ -631,6 +643,8 @@ def field_singleton_schema(  # noqa: C901 (ignore complexity)
         )
     if field.type_ is Any:
         return {}, definitions  # no restrictions
+    if is_callable_type(field.type_):
+        raise SkipField(f'Callable {field.name} was excluded from schema since JSON schema has no equivalent type.')
     f_schema: Dict[str, Any] = {}
     if issubclass(field.type_, Enum):
         f_schema.update({'enum': [item.value for item in field.type_]})  # type: ignore
@@ -698,7 +712,7 @@ def get_annotation_from_schema(annotation: Any, schema: Schema) -> Type[Any]:
             annotation, (ConstrainedInt, ConstrainedFloat, ConstrainedDecimal, bool)
         ):
             # Is numeric type
-            attrs = ('gt', 'lt', 'ge', 'le')
+            attrs = ('gt', 'lt', 'ge', 'le', 'multiple_of')
             numeric_type = next(t for t in numeric_types if issubclass(annotation, t))  # pragma: no branch
             constraint_func = _map_types_constraint[numeric_type]
 
@@ -712,3 +726,12 @@ def get_annotation_from_schema(annotation: Any, schema: Schema) -> Type[Any]:
                 constraint_func = cast(Callable[..., type], constraint_func)
                 return constraint_func(**kwargs)
     return annotation
+
+
+class SkipField(Exception):
+    """
+    Utility exception used to exclude fields from schema.
+    """
+
+    def __init__(self, message):
+        self.message = message