+++ pydantic/utils.py
@@ -29,6 +29,11 @@ except ImportError:
 if TYPE_CHECKING:  # pragma: no cover
     from . import main  # noqa
 
+if sys.version_info < (3, 7):
+    from typing import Callable
+else:
+    from collections.abc import Callable
+
 PRETTY_REGEX = re.compile(r'([\w ]*?) *<(.*)> *')
 
 
@@ -248,3 +253,7 @@ def resolve_annotations(raw_annotations, module):
             pass
         annotations[name] = value
     return annotations
+
+
+def is_callable_type(type_):
+    return type_ is Callable or getattr(type_, '__origin__', None) is Callable